tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/before_pseudo.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - :before pseudo-element</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-name-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="Animation properties apply to :before">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes grow {
         from {
            width: 100px;
         }
         to {
            width: 200px;
         }
      }
      div:before {
         content: "";
         display: block;
         width: 100px;
         height: 100px;
         background: blue;
      }
      div.test:before {
         animation-name: grow;
         animation-duration: 0.7s;
         animation-fill-mode: forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         setTimeout(function() {
            var passed = (getComputedStyle(div, ':before').width == '200px');
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
            test(function(){assert_true(passed);}, testName);
            done();
         }, 1000);
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The block below should gradually grow twice as wide.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
